人像克隆修复非人像照片标记

+++ frameworks/services/media_backup_extension/src/analysis_album/beauty_score_clone.cpp
@@ -196,7 +196,7 @@ void BeautyScoreClone::BatchInsertBeautyScores(
     }
 
     migrateScoreNum_ += rowNum;
-    migrateScoreFileNumber_ += fileIdSet.size();
+    migrateScoreFileNumber_ += static_cast<int64_t>(fileIdSet.size());
 }
 
 NativeRdb::ValuesBucket BeautyScoreClone::CreateValuesBucketFromBeautyScoreTbl(
